dynamically enable/disable fwapc in assets and service

+++ scripts/prep/17_add_fwapc.sh
@@ -5,9 +5,13 @@
 source ${FIREWALLA_HOME}/platform/platform.sh
 
 if [[ ${HAVE_FWAPC} == "yes" ]]; then
-  sudo cp ${FIREWALLA_HOME}/etc/fwapc.service /etc/systemd/system/.
-  sudo systemctl daemon-reload
-  test -f /home/pi/.firewalla/run/assets/fwapc && sudo systemctl start fwapc &>/dev/null # fwapc will be downloaded and updated via assets framework
+  if ! cmp -s ${FIREWALLA_HOME}/etc/fwapc.service /etc/systemd/system/fwapc.service; then
+    sudo cp ${FIREWALLA_HOME}/etc/fwapc.service /etc/systemd/system/.
+    sudo systemctl daemon-reload
+    if systemctl -q is-active fwapc; then
+      sudo systemctl restart fwapc
+    fi
+  fi
 fi
 
 exit 0
